all-ns-all-ips - add cli flags and refactor out the extract nameserver logic
